Split des sources de l'application
