+++ app/common.cpp
@@ -1,0 +1,2 @@
+#include <subs/maths/subs_fparser.h>
+#include <subs/maths/subs_exprtk.h>
